implement /proc parsing
Parse all relevant files in /proc and /dev to fully replace
github.com/shirou/gopsutil.  Also add unit tests, which are now
possible as we can test our parsers.

Nothing has changed with regards to the API as containers/psgo is
currently only meant to replace ps (1) and not to offer a generic
library to parse /proc and /dev/{tty*,pts/*}.  From a technical
point of view, all code is in containers/psgo/pg to avoid exposing
any public API.

Performance-wise, we are around twice as fast as before as we
pre-load all data and do not load on-demand which is costly.
Furthermore, we now correctly treat tty devices and have a
consistent way of dealing with race conditions.

Note that the parses extract more data from /proc/$pid/stat and
from /proc/$pid/status than necessary.  This way, we can easily
extend the format descriptors to return more information.
Otherwise, the parsing of redundant fields can easily be removed.

Fixes: #4
Fixes: #7
Signed-off-by: Valentin Rothberg <vrothberg@suse.com>