add `JoinNamespaceAndProcessInfoByPids` API
Add a new API, `JoinNamespaceAndProcessInfoByPids`, which is similar to
`JoinNamespaceAndProcessInfo` but takes a slice of pids as an argument.
To avoid duplicate entries (e.g., when two or more containers share the
same PID namespace), a given PID namespace will be joined only once.

Notice that the CLI of the sample implementation in `./bin/psgo` has
changed.  The `-pid` flag has been removed in favor of the`-pids` flag.
A new `-join` bool flag controls if the namespaces of the provided pids
shall be joined.

Signed-off-by: Valentin Rothberg <vrothberg@suse.com>